+++ CSharpSourceCode/TOR_Core.csproj
@@ -529,6 +529,7 @@
     <Compile Include="CampaignMechanics\WaaaghMeter\WaaaghMeterVM.cs" />
     <Compile Include="CampaignMechanics\WaaaghMeter\WaaaghMeterMapView.cs" />
     <Compile Include="CampaignMechanics\WaaaghMeter\WaaaghMeterGlobalLayer.cs" />
+    <Compile Include="CampaignMechanics\WaaaghMeter\WaaaghHelper.cs" />
     <Compile Include="CampaignMechanics\CustomResources\ChivalryHelper.cs" />
     <Compile Include="CampaignMechanics\CustomResources\CustomResource.cs" />
     <Compile Include="CampaignMechanics\CustomResources\CustomResourceManager.cs" />
