+++ GUI/Prefabs/Map/WaaaghMeter.xml
@@ -9,68 +9,68 @@
       <Children>
         <!-- Background -->
         <Widget DoNotAcceptEvents="true" WidthSizePolicy="Fixed" HeightSizePolicy="StretchToParent"
-                SuggestedWidth="20" HorizontalAlignment="Center"
+                SuggestedWidth="10" HorizontalAlignment="Center"
                 Sprite="BlankWhiteSquare_9" Color="#1A1A1AFF" />
 
         <!-- Progress Fill (red, from bottom) -->
         <Widget DoNotAcceptEvents="true" WidthSizePolicy="Fixed" HeightSizePolicy="Fixed"
-                SuggestedWidth="20" SuggestedHeight="@FillHeight"
+                SuggestedWidth="10" SuggestedHeight="@FillHeight"
                 HorizontalAlignment="Center" VerticalAlignment="Bottom"
                 Sprite="BlankWhiteSquare_9" Color="#FF0000FF" />
+
+        <!-- Border (rendered before icons so icons appear on top) -->
+        <Widget DoNotAcceptEvents="true" WidthSizePolicy="Fixed" HeightSizePolicy="StretchToParent"
+                SuggestedWidth="10" HorizontalAlignment="Center"
+                Sprite="BlankWhiteSquare_9" Color="#50463C44" />
 
         <!-- Level Icons (4 icons placed at level thresholds) with hover tooltips -->
         <!-- Level 0 icon: Bottom (0%) - Internal Fightin' -->
         <Widget WidthSizePolicy="Fixed" HeightSizePolicy="Fixed" SuggestedWidth="43" SuggestedHeight="43"
                 HorizontalAlignment="Center" VerticalAlignment="Bottom" MarginBottom="-11"
-                Sprite="BlankWhiteSquare_9" Color="#00000040">
+                Sprite="BlankWhiteSquare_9" Color="#00000000">
           <Children>
             <Widget DoNotAcceptEvents="true" WidthSizePolicy="Fixed" HeightSizePolicy="Fixed" SuggestedWidth="32" SuggestedHeight="32"
                     HorizontalAlignment="Center" VerticalAlignment="Center"
-                    Sprite="shyish" />
+                    Sprite="waagh1_icon" />
             <HintWidget DataSource="{Level0Hint}" DoNotAcceptEvents="true" WidthSizePolicy="StretchToParent" HeightSizePolicy="StretchToParent" Command.HoverBegin="ExecuteBeginHint" Command.HoverEnd="ExecuteEndHint" />
           </Children>
         </Widget>
 
-        <!-- Level 1 icon: ~33% - Petty Squabblin' -->
+        <!-- Level 1 icon: 25% (250/1000) - Petty Squabblin' -->
         <Widget WidthSizePolicy="Fixed" HeightSizePolicy="Fixed" SuggestedWidth="43" SuggestedHeight="43"
-                HorizontalAlignment="Center" VerticalAlignment="Bottom" MarginBottom="94"
-                Sprite="BlankWhiteSquare_9" Color="#00000040">
+                HorizontalAlignment="Center" VerticalAlignment="Bottom" MarginBottom="66"
+                Sprite="BlankWhiteSquare_9" Color="#00000000">
           <Children>
             <Widget DoNotAcceptEvents="true" WidthSizePolicy="Fixed" HeightSizePolicy="Fixed" SuggestedWidth="32" SuggestedHeight="32"
                     HorizontalAlignment="Center" VerticalAlignment="Center"
-                    Sprite="ghur" />
+                    Sprite="waagh2_icon" />
             <HintWidget DataSource="{Level1Hint}" DoNotAcceptEvents="true" WidthSizePolicy="StretchToParent" HeightSizePolicy="StretchToParent" Command.HoverBegin="ExecuteBeginHint" Command.HoverEnd="ExecuteEndHint" />
           </Children>
         </Widget>
 
-        <!-- Level 2 icon: ~66% - 'Ere We Go! -->
+        <!-- Level 2 icon: 60% (600/1000) - 'Ere We Go! -->
         <Widget WidthSizePolicy="Fixed" HeightSizePolicy="Fixed" SuggestedWidth="43" SuggestedHeight="43"
-                HorizontalAlignment="Center" VerticalAlignment="Bottom" MarginBottom="211"
-                Sprite="BlankWhiteSquare_9" Color="#00000040">
+                HorizontalAlignment="Center" VerticalAlignment="Bottom" MarginBottom="189"
+                Sprite="BlankWhiteSquare_9" Color="#00000000">
           <Children>
             <Widget DoNotAcceptEvents="true" WidthSizePolicy="Fixed" HeightSizePolicy="Fixed" SuggestedWidth="32" SuggestedHeight="32"
                     HorizontalAlignment="Center" VerticalAlignment="Center"
-                    Sprite="chamon" />
+                    Sprite="waagh3_icon" />
             <HintWidget DataSource="{Level2Hint}" DoNotAcceptEvents="true" WidthSizePolicy="StretchToParent" HeightSizePolicy="StretchToParent" Command.HoverBegin="ExecuteBeginHint" Command.HoverEnd="ExecuteEndHint" />
           </Children>
         </Widget>
 
-        <!-- Level 3 icon: Top (100%) - WAAAGH!!!! -->
+        <!-- Level 3 icon: 90% (900/1000) - WAAAGH!!!! -->
         <Widget WidthSizePolicy="Fixed" HeightSizePolicy="Fixed" SuggestedWidth="43" SuggestedHeight="43"
-                HorizontalAlignment="Center" VerticalAlignment="Bottom" MarginBottom="318"
-                Sprite="BlankWhiteSquare_9" Color="#00000040">
+                HorizontalAlignment="Center" VerticalAlignment="Bottom" MarginBottom="294"
+                Sprite="BlankWhiteSquare_9" Color="#00000000">
           <Children>
             <Widget DoNotAcceptEvents="true" WidthSizePolicy="Fixed" HeightSizePolicy="Fixed" SuggestedWidth="32" SuggestedHeight="32"
                     HorizontalAlignment="Center" VerticalAlignment="Center"
-                    Sprite="aqshy" />
+                    Sprite="waagh4_icon" />
             <HintWidget DataSource="{Level3Hint}" DoNotAcceptEvents="true" WidthSizePolicy="StretchToParent" HeightSizePolicy="StretchToParent" Command.HoverBegin="ExecuteBeginHint" Command.HoverEnd="ExecuteEndHint" />
           </Children>
         </Widget>
-
-        <!-- Border -->
-        <Widget DoNotAcceptEvents="true" WidthSizePolicy="Fixed" HeightSizePolicy="StretchToParent"
-                SuggestedWidth="20" HorizontalAlignment="Center"
-                Sprite="BlankWhiteSquare_9" Color="#50463C44" />
       </Children>
     </Widget>
   </Window>
